(#236) refactor: 풀업뷰(카드뷰) 를 뷰에서 컨트롤러 형태로 수정
